Updates from @uboness review: int counts changed to longs, asciidoc examples changed to lowercase, base class change to SignificantTerms, code formatting, parser parses “format” field. I’ve added a “TODO” comment for the refactoring suggestion here: https://github.com/elasticsearch/elasticsearch/pull/5146#discussion_r10500246 - as this should be considered as part of future changes

+++ src/main/java/org/elasticsearch/search/aggregations/bucket/significant/SignificantLongTermsAggregator.java
@@ -43,7 +43,7 @@ public class SignificantLongTermsAggregator extends LongTermsAggregator {
         this.termsAggFactory = termsAggFactory;
     }
 
-    protected int numCollectedDocs;
+    protected long numCollectedDocs;
     private SignificantTermsAggregatorFactory termsAggFactory;
 
     @Override
@@ -60,8 +60,8 @@ public class SignificantLongTermsAggregator extends LongTermsAggregator {
 
         ContextIndexSearcher searcher = context.searchContext().searcher();
         IndexReader topReader = searcher.getIndexReader();
-        int supersetSize = topReader.numDocs();
-        int subsetSize = numCollectedDocs;
+        long supersetSize = topReader.numDocs();
+        long subsetSize = numCollectedDocs;
 
         BucketSignificancePriorityQueue ordered = new BucketSignificancePriorityQueue(size);
         SignificantLongTerms.Bucket spare = null;
